fix(context-menu): focus the context menu after the position is set
This fixes the jump of the table when the context menu was opened
in a scrolled table. Previously, the focus was set on a context menu
positioned in the top left corner of the table which forced the browser
to scroll the position to the top left corner. Then, the position
of the context menu was correctly set, but the scroll position
was already changed.

Now, the context menu is positioned correctly first and only then
can it be focused.

Related to https://github.com/camunda/camunda-modeler/issues/1433